pyproject.toml
[project]
name = "scrape-embed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.66.3",
    "psycopg>=3.2.6",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
]
